split preview panel and widgets code (drag+rotate)
